adding collection to not run in parallel
